Lift the state
